requirements.txt
Flask
numpy>=1.19.0
torch>=2.2.0
torchvision>=0.17.0
facenet-pytorch>=2.5.2
Pillow>=8.0.0
opencv-python>=4.5.0
matplotlib>=3.3.0
werkzeug>=2.0.0
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
pytesseract==0.3.10
moviepy==1.0.3
pydub==0.25.1
transformers==4.41.2
scikit-learn>=0.24.0
seaborn>=0.11.0
kraken==4.3.13
scikit-image==0.23.2